utilize normalize-task in pending page as well
